refactor(server): page layout tab
