+++ packages/twenty-server/src/engine/metadata-modules/flat-page-layout-tab/utils/from-delete-page-layout-tab-input-to-flat-page-layout-tab-or-throw.util.ts
@@ -9,7 +9,7 @@ import { type FlatPageLayoutTab } from 'src/engine/metadata-modules/flat-page-la
 import {
   PageLayoutTabException,
   PageLayoutTabExceptionCode,
-} from 'src/engine/metadata-modules/page-layout/exceptions/page-layout-tab.exception';
+} from 'src/engine/metadata-modules/page-layout-tab/exceptions/page-layout-tab.exception';
 
 export type DeletePageLayoutTabInput = {
   id: string;
